ehmkah: refactoring simpleformatasciidocaction as a base class for bold, italic, monospace formating https://github.com/asciidoctor/asciidoctor-intellij-plugin/issues/86
