+++ src/main/java/org/asciidoc/intellij/actions/asciidoc/FormatAsciiDocAction.java
@@ -70,30 +70,4 @@ public abstract class FormatAsciiDocAction extends AsciiDocAction {
     });
   }
 
-  protected String updateSelectionIntern(String selection, String symbol) {
-    if (containsSymbol(selection, symbol)) {
-      return removeSymbol(selection, symbol);
-    }
-    return appendSymbol(selection, symbol);
-  }
-
-  private String appendSymbol(String selection, String symbol) {
-    String doubleSymbol = symbol + symbol;
-    return doubleSymbol + selection + doubleSymbol;
-  }
-
-  private String removeSymbol(String selection, String symbol) {
-    if (selection.startsWith(symbol + symbol)) {
-      return selection.substring(2, selection.length() - 2);
-    }
-    return selection.substring(1, selection.length() - 1);
-  }
-
-  private boolean containsSymbol(String selection, String symbol) {
-    String doubleSymbol = symbol + symbol;
-    return (selection.startsWith(symbol) && selection.endsWith(symbol)) ||
-        (selection.startsWith(doubleSymbol) && selection.endsWith(doubleSymbol));
-  }
-
-
 }
